requirements.txt
amqp
asgiref
async-timeout
billiard
celery
certifi
charset-normalizer
click
click-didyoumean
click-plugins
click-repl
coreapi
coreschema
Deprecated
dj-database-url
Django
django-cors-headers
django-filter
djangorestframework
djangorestframework-simplejwt
drf-yasg
idna
inflection
itypes
Jinja2
kombu
MarkupSafe
packaging
Pillow
prompt-toolkit
psycopg2-binary
PyJWT
pyparsing
python-decouple
pytz
redis
requests
ruamel.yaml
ruamel.yaml.clib
six
sqlparse
uritemplate
urllib3
vine
wcwidth
wrapt
